Initial push of mcd2c because Alex asked me to
The code here is very broken, it mostly consists of two files, mcd2c.py
and mcd2c_2.py. The rest of the files in the repo are C data utilities
for serializing and deserializing minecraft data.

mcd2c.py is an old attempt at writing a C code generator for the
ProtoDef format, it's based on the cfile repo by cogu found here:
https://github.com/cogu/cfile

I keep it in the repo because it can be a useful reference while writing
mcd2c_2.py.

mcd2c_2.py is my new, clean attempt at the code generator. So far it
supports generating the packet structs and 'walk' functions. Once it
supports 'size', 'dec', 'enc', and 'free' we'll be off to the races.

Everything here is very in-development, very rough, I wouldn't normally
push this code but Alex wanted to see what I was working on.